webapi/tct-calendar-tizen-tests/calendar/Calendar_CalendarAlarm_constructor_TimeDuration.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
this list of conditions and the following disclaimer in the documentation
and/or other materials provided with the distribution.
Neither the name of Intel Corporation nor the names of its contributors
may be used to endorse or promote products derived from this work without
specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Jenny Cao <jenny.q.cao@intel.com>
        Yanan Xu <yananx.xu@intel.com>

-->

<html>
<head>
<title>Calendar_CalendarAlarm_constructor_TimeDuration</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/calendar_common.js"></script>
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
//==== TEST: Calendar_CalendarAlarm_constructor_TimeDuration
//==== LABEL Check CalendarAlarm constructor: TimeDuration
//==== PRIORITY P2
//==== SPEC Tizen Web API:Social:Calendar:CalendarAlarm:constructor M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/calendar.html
//==== TEST_CRITERIA CONSTRM CONSTRA

test(function () {
    var before = new tizen.TimeDuration(15, "MINS"),
        alarm = new tizen.CalendarAlarm(before, "DISPLAY", "description of DISPLAY");

    assert_equals(alarm.absoluteDate, null, "Incorrect absoluteDate");
    assert_true(alarm.before.equalsTo(before), "Incorrect before");
    assert_true(alarm.method in AlarmMethod, "Incorrect method");
    assert_not_equals(alarm.description, null, "Incorrect description");
}, document.title);

</script>
</body>
</html>
